refactor: move is_ignored and status fields from core to connect_twilio
These fields are Twilio-specific: is_ignored controls Twilio number sync,
status tracks CallerID validation state. Removed is_ignored filter from
FreeSWITCH DID lookup as it's not relevant there.
