+++ connect/views/number_views.xml
@@ -10,7 +10,6 @@
                 <field name="destination"/>
                 <field name="user"/>
                 <field name="callflow"/>
-                <field name="is_ignored"/>
             </list>
         </field>
     </record>
@@ -25,7 +24,6 @@
                         <group string="Number">
                             <field name="phone_number"/>
                             <field name="friendly_name"/>
-                            <field name="is_ignored"/>
                             <field name="is_default"/>
                         </group>
                         <group string="Routing">
