+++ connect/views/outgoing_callerid_views.xml
@@ -7,7 +7,6 @@
             <list>
                 <field name="number"/>
                 <field name="friendly_name"/>
-                <field name="status"/>
                 <field name="callerid_type"/>
                 <field name="is_default"/>
             </list>
@@ -25,7 +24,6 @@
                             <field name="number"/>
                             <field name="friendly_name"/>
                             <field name="callerid_type"/>
-                            <field name="status" readonly="1"/>
                             <field name="is_default"/>
                         </group>
                         <group string="Users">
